# math-service-openapi-spec.yaml
openapi: 3.0.1
info:
  title: Math Service
  description: Simple authorized math service
  termsOfService: replace_with_url
  contact:
    email: mathservice@dmgt.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 0.0.1
...
paths:
...
  /sub:
    post:
      'x-swagger-router-controller': 'endpoint_subtract'
      operationId: subtract
      summary: Subtract two numbers from each other.
      security:
        - bearerSelfAuth: []
      parameters:
        - name: authorization
          schema:
            type: string
          required: true
          in: header
          description: |-
            The client must present an Usher-issued JWT with the required scope.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                left:
                  type: number
                right:
                  type: number
              required:
                - left
                - right
      responses:
        '200':
          description: Returns a difference
        '403':
          description: If unauthorized.
...
components:
  securitySchemes:
    bearerSelfAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT